EncrytionAndDecryption
This program will use a very simple algorithm to encrypt a string entered by a user. 
For a string e.g 'abcd', it will move each character up one space in the dictionary and then when it has done that it will print the string in reverse order ('abcd --> 'edcb').